==> R/00_all.qmd <==
---
title: "00_all"
output:
  html:
    self-contained: true
    output_dir: "./results"
    file_name: "00_all.html"
    resources: false
editor: visual
---

## Loading libraries

```{r}
library(tidyverse)
library(dplyr)
library(patchwork)
library(ggplot2)
library(knitr)
library("quarto")
library(here)
library(broom)
library(cowplot)
library(ggrepel)
library(readr)
library(stringr)
```

Create results folder if non exisisting

```{r}
# Create results folder if non-existing
if (!dir.exists("../results")) {
  dir.create("../results")
}

```

We will use quarto::render

## Loading data

```{r include=FALSE}

# Render the other .qmd files
if (file.exists("R/01_load.qmd")) {
  quarto::quarto_render("R/01_load.qmd")
} else {
  quarto::quarto_render("../R/01_load.qmd")
}
#move to results
file.rename(here("R/01_load.html"), here("results/01_load.html"))


```

## clean data

```{r}

if (file.exists("R/02_clean.qmd")) {
  quarto::quarto_render("R/02_clean.qmd")
} else {
  quarto::quarto_render("../R/02_clean.qmd")
}
#move to results:
file.rename(here("R/02_clean.html"), here("results/02_clean.html"))

```

## Augment the data

```{r}

if (file.exists("R/03_augment.qmd")) {
  quarto::quarto_render("R/03_augment.qmd")
} else {
  quarto::quarto_render("../R/03_augment.qmd")
}
#move to results:
file.rename(here("R/03_augment.html"), here("results/03_augment.html"))


```

## Describe the data

```{r}

if (file.exists("R/04_describe.qmd")) {
  quarto::quarto_render("R/04_describe.qmd")
} else {
  quarto::quarto_render("../R/04_describe.qmd")
}
#mnove to results
file.rename(here("R/04_describe.html"), here("results/04_describe.html"))


```

## Analysis 1

```{r}

if (file.exists("R/05_analysis_1.qmd")) {
  quarto::quarto_render("R/05_analysis_1.qmd")
} else {
  quarto::quarto_render("../R/05_analysis_1.qmd")
}

#move to results:
file.rename(here("R/05_analysis_1.html"), here("results/05_analysis_1.html"))

```

## Analysis 2 PCA

```{r}

if (file.exists("R/06_analysis_2_PCA.qmd")) {
  quarto::quarto_render("R/06_analysis_2_PCA.qmd")
} else {
  quarto::quarto_render("../R/06_analysis_2_PCA.qmd")
}
file.rename(here("R/06_analysis_2_PCA.html"), here("results/06_analysis_2_PCA.html"))


```

## For all

```{r echo=FALSE}
if (dir.exists("results")) {
  html_files <- list.files("results", pattern = "\\.html$", full.names = TRUE)
} else if (dir.exists("../results")) {
  html_files <- list.files("../results", pattern = "\\.html$", full.names = TRUE)
} else {
  stop("Neither 'results' nor '../results' directory found.")
}

# Read and combine all HTML content safely
html_content <- sapply(html_files, function(file) {
  paste(readLines(file, encoding = "UTF-8", warn = FALSE), collapse = "\n")
}, USE.NAMES = FALSE)

# Combine all contents into a single string
combined_html <- paste(html_content, collapse = "\n")

# Write the combined HTML to the 00_all.html file
if (dir.exists("results")) {
  cat(combined_html, file = "results/00_all.html", sep = "\n")
} else {
  cat(combined_html, file = "../results/00_all.html", sep = "\n")
}



```
