==> R/04_describe.qmd <==
---
title: "04_describe"
format:
  html:
    self-contained: true
editor: visual
---

## Loading libraries

```{r}
library(tidyverse)
```

## Loading data

```{r}
data_aug <- read_csv("../data/03_data_aug.csv")
```

## Descriptive statistics

Number of subjects

```{r}
nr_of_subjects <- data_aug %>%
  summarise(distinct_ID_count = n_distinct(ID))

nr_of_subjects
```

### Dimensions

Rows is the number of data points for STDs, so if a subject has 0 or 1 STD they will have 1 row. If a subject has multiple STDs, they will have one row per STD diagnosis.

```{r}
data_aug |> 
  group_by(ID) |> 
  dim()
```

Age distribution

```{r}
data_aug |> 
  group_by(ID) |> 
  ggplot(aes(x = Age)) + 
  geom_boxplot()
```

Number of Cancer positives

```{r}
data_aug |> 
  group_by(Dx_Cancer) |> 
  count()
```

### Smoking

Number of smokers

```{r}
data_aug |> 
  group_by(Smokes) |> 
  count()
```

Distribution of years of smoking

```{r}
data_aug |> 
  group_by(ID) |> 
  ggplot(aes(x = Smokes_years)) + 
  geom_boxplot()
```

### Sexual health

First sexual intercourse distribution

```{r}
data_aug |> 
  group_by(ID) |> 
  ggplot(aes(x = First_sexual_intercourse)) + 
  geom_boxplot()
```

Number of pregnancies

```{r}
data_aug |> 
  group_by(ID) |> 
  ggplot(aes(x = Num_of_pregnancies)) + 
  geom_boxplot()
```

Number of sexual partners

```{r}
data_aug |> 
  group_by(ID) |> 
  ggplot(aes(x = Number_of_sexual_partners)) + 
  geom_boxplot()
```
