==> R/06_analysis_2_PCA.qmd <==
---
title: "PCA"
format:
  html:
    self-contained: true
editor: visual
---

## Libraries

```{r}
library(tidyverse)
library(broom)
library(cowplot)
library(ggrepel)
```

## Loading

```{r}
data_aug <- read.csv("../data/03_data_aug.csv")
```

## Making a PCA plot based on patients have cancer or not. Created a new column based on Dx_cancer

```{r}
cancer_data <- data_aug |> 
  select(-Time_since_first_diagnosis, -Time_since_last_diagnosis, Dx_CIN) |> drop_na() 

pca_fit_cancer <- cancer_data |> 
  select(where(is.numeric)) |>  
  prcomp(scale = TRUE)

scatterplot_pca <- pca_fit_cancer |> 
  augment(cancer_data) |> 
  ggplot(aes(.fittedPC1, .fittedPC2, color = factor(Cancer))) + 
  geom_point(size = 1.5) +
  theme_half_open(12) + background_grid() + scale_color_discrete(labels = c("0" = "Negative", "1" = "Positive")) +
  labs(color = "Cervical Cancer Present") + ggtitle("PCA Plot") + theme(plot.title = element_text(hjust = 0.5))

summary(pca_fit_cancer)

scatterplot_pca

#save plot
ggsave("../results/images/06_scatterplot_pca.png", plot = scatterplot_pca, bg = "white")

```

The negative patients form a dense cluster in the right corner, whereas it is more spread out to the lower-left corner for the positive ones. Patients with positive cancer diagnosis have more distinct patterns in their values compared to those with no cancer. This separation suggests that features e.g., age, smoking or STD's might contribute strongly to distinguish cancer-positive patients.

### The data in PC coordinates

```{r}
pca_fit_cancer |> 
  tidy(matrix = "rotation")
# define arrow style for plotting
arrow_style <- arrow(
  angle = 20, ends = "first", type = "closed", length = grid::unit(10, "pt")
)


pc_coordinates <- pca_fit_cancer |>
  tidy(matrix = "rotation") %>%
  pivot_wider(names_from = "PC", names_prefix = "PC", values_from = "value") %>%
  ggplot(aes(PC1, PC2)) +
  geom_segment(xend = 0, yend = 0, arrow = arrow_style) +
  geom_text_repel(
    aes(label = column),
    color = "darkblue"
  ) +
  #xlim(-0.5, .3) + ylim(-.25, 0.5) +
  coord_fixed() +
  theme_minimal_grid(12) +
  ggtitle("The data in the PC coordinates") + 
  theme(plot.title = element_text(hjust = 0.5))


#save plot
ggsave("../results/images/06_pc_coordinates.png", plot = pc_coordinates, bg = "white")
```

The arrows shows how the features contribute to the principal components analysis.

### The variance explained by each PC

A general rule is to keep the components that explains up to 95% of the variance.

```{r}
pca_fit_cancer |>
  tidy(matrix = "eigenvalues")

pc_variance <- pca_fit_cancer |>
  tidy(matrix = "eigenvalues") |>
  ggplot(aes(PC, percent)) +
  geom_col(fill = "darkblue", alpha = 0.8) +
  scale_x_continuous(breaks = 1:17) +
  scale_y_continuous(
    labels = scales::percent_format(),
    expand = expansion(mult = c(0, 0.01))
  ) +
  theme_minimal_hgrid(12) +
  ggtitle("The variance explained by each PC") + theme(plot.title = element_text(hjust = 0.5))

#save plot
ggsave("../results/images/06_pc_variance.png", plot = pc_variance, bg = "white")
```

After PC4 and PC% the variance explained by each PC drops significantly forming an 'elbow'. After PC10 the other components explain very little variance. These components might not be useful for analyzing or modeling.
